pyproject.toml
[project]
name = "socialnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==4.2.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv==1.0.1",
]
